refactor: split parsing logic from code runner logic. Move defaultPrompt to the main which is the only place it's used
